+++ src/PostgREST/Request/QueryParams.hs
@@ -44,16 +44,18 @@ import PostgREST.RangeQuery              (NonnegRange, allRange,
                                           rangeGeq, rangeLimit,
                                           rangeOffset, restrictRange)
 
-import PostgREST.Request.Types (EmbedParam (..), EmbedPath, Field,
-                                Filter (..), FtsOperator (..),
-                                JoinType (..), JsonOperand (..),
-                                JsonOperation (..), JsonPath, ListVal,
-                                LogicOperator (..), LogicTree (..),
-                                OpExpr (..), Operation (..),
-                                OrderDirection (..), OrderNulls (..),
-                                OrderTerm (..), QPError (..),
-                                SelectItem, SimpleOperator (..),
-                                SingleVal, TrileanVal (..))
+import PostgREST.Request.ReadQuery (SelectItem)
+import PostgREST.Request.Types     (EmbedParam (..), EmbedPath, Field,
+                                    Filter (..), FtsOperator (..),
+                                    JoinType (..), JsonOperand (..),
+                                    JsonOperation (..), JsonPath,
+                                    ListVal, LogicOperator (..),
+                                    LogicTree (..), OpExpr (..),
+                                    Operation (..),
+                                    OrderDirection (..),
+                                    OrderNulls (..), OrderTerm (..),
+                                    QPError (..), SimpleOperator (..),
+                                    SingleVal, TrileanVal (..))
 
 import Protolude hiding (try)
